Split aria-pressed helpers and URL serialization out of punch_controller
The Stimulus controller was carrying two self-contained concerns that
have nothing to do with Stimulus: the aria-pressed read/write helpers
and the punchcard URL state schema. Extracted both to
app/javascript/utils/ so the controller only owns DOM indexing,
intent capture, and event wiring.

Imports use the importmap "utils/<name>" pin (configured in
config/importmap.rb). Controller dropped from 6580 → 5370
non-whitespace chars; full suite (350) green.
